# Implement a more general slice pool