--- a/client/client.go
+++ b/client/client.go
@@ -8,8 +8,15 @@ import (
 
 	onet "github.com/Jigsaw-Code/outline-ss-server/net"
 	ss "github.com/Jigsaw-Code/outline-ss-server/shadowsocks"
+	"github.com/Jigsaw-Code/outline-ss-server/slicepool"
 	"github.com/shadowsocks/go-shadowsocks2/socks"
 )
+
+// clientUDPBufferSize is the maximum supported UDP packet size in bytes.
+const clientUDPBufferSize = 16 * 1024
+
+// udpPool stores the byte slices used for storing encrypted packets.
+var udpPool = slicepool.MakePool(clientUDPBufferSize)
 
 // Client is a client for Shadowsocks TCP and UDP connections.
 type Client interface {
@@ -102,8 +109,9 @@ func (c *packetConn) WriteTo(b []byte, addr net.Addr) (int, error) {
 	if socksTargetAddr == nil {
 		return 0, errors.New("Failed to parse target address")
 	}
-	cipherBuf := newUDPBuffer()
-	defer freeUDPBuffer(cipherBuf)
+	box := udpPool.Slice()
+	cipherBuf := box.Acquire()
+	defer box.Release()
 	saltSize := c.cipher.SaltSize()
 	// Copy the SOCKS target address and payload, reserving space for the generated salt to avoid
 	// partially overlapping the plaintext and cipher slices since `Pack` skips the salt when calling
@@ -119,8 +127,9 @@ func (c *packetConn) WriteTo(b []byte, addr net.Addr) (int, error) {
 
 // ReadFrom reads from the embedded PacketConn and decrypts into `b`.
 func (c *packetConn) ReadFrom(b []byte) (int, net.Addr, error) {
-	cipherBuf := newUDPBuffer()
-	defer freeUDPBuffer(cipherBuf)
+	box := udpPool.Slice()
+	cipherBuf := box.Acquire()
+	defer box.Release()
 	n, err := c.UDPConn.Read(cipherBuf)
 	if err != nil {
 		return 0, nil, err