This saves one allocation on every UDP packet on the client, by
storing pointers to slices instead of slice values:

```
Master:
BenchmarkShadowsocksClient_UDPWrite-4   	   10000	    235685 ns/op	    1329 B/op	      22 allocs/op
BenchmarkShadowsocksClient_ListenUDP-4   	   15590	     76784 ns/op	    5260 B/op	      79 allocs/op
```

```
This branch:
BenchmarkShadowsocksClient_UDPWrite-4   	   10000	    194026 ns/op	    1296 B/op	      21 allocs/op
BenchmarkShadowsocksClient_ListenUDP-4   	   14834	     77452 ns/op	    5191 B/op	      77 allocs/op
```

More importantly, this is a prerequisite for larger savings by using
buffer pools for TCP streams.