@bemasc 2020-10-09
LazySlice does seem like a better name.

Thanks to garbage collection, returning a pointer to a local variable works
fine.  The weird construction is because casting a *[]byte to interface{}
is free, but casting a []byte to interface{} requires allocating a
temporary interface{}.

On Fri, Oct 9, 2020, 1:06 PM Vinicius Fortuna <notifications@github.com>
wrote:

> *@fortuna* requested changes on this pull request.
> ------------------------------
>
> In client/client.go
> <https://github.com/Jigsaw-Code/outline-ss-server/pull/89#discussion_r502563532>
> :
>
> > @@ -102,8 +109,9 @@ func (c *packetConn) WriteTo(b []byte, addr net.Addr) (int, error) {
>  	if socksTargetAddr == nil {
>  		return 0, errors.New("Failed to parse target address")
>  	}
> -	cipherBuf := newUDPBuffer()
> -	defer freeUDPBuffer(cipherBuf)
> +	box := udpPool.Slice()
>
> Perhaps we call this something like LazySlice, to not confuse with a
> standard slice?
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/Jigsaw-Code/outline-ss-server/pull/89#pullrequestreview-505850464>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AABO3SKVDB3VU5WHBVTHRGTSJ47JDANCNFSM4SKH4XPQ>
> .
>